pacbio and ont

+++ example/example.sh
@@ -143,35 +143,46 @@ then
 	python ${BASEDIR}/../scripts/merge.py *.rna.illumina.pe.json.gz | gzip -c > rna.illumina.pe.ms.json.gz
 	rm *.rna.illumina.pe.json.gz
     fi
-    exit;
+
     # PacBio
-    for SAMPLE in NA19238
-    do
-	if [ ! -f ${SAMPLE}.pacbio.bam ]
-	then
-	    wget "ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/data_collections/hgsv_sv_discovery/working/20180102_pacbio_blasr_reheader/${SAMPLE}.pacbio-blasr-grch38-reheader.20180102.chr*.bam"
-	    for F in ${SAMPLE}.pacbio-blasr-grch38-reheader.20180102.*.bam
-	    do
-		if [ ! -f ${F}.bai ]
-		then
-		    samtools index ${F}
-		fi
-	    done
-	    samtools merge ${SAMPLE}.pacbio.bam ${SAMPLE}.pacbio-blasr-grch38-reheader.20180102.*.bam
-	    samtools index ${SAMPLE}.pacbio.bam
-	    rm ${SAMPLE}.pacbio-blasr-grch38-reheader.20180102.chr*
-	fi
-    done
+    if [ ! -f dna.pacbio.se.ms.json.gz ]
+    then
+	for SAMPLE in NA19238
+	do
+	    if [ ! -f ${SAMPLE}.pacbio.bam ]
+	    then
+		wget "ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/data_collections/hgsv_sv_discovery/working/20180102_pacbio_blasr_reheader/${SAMPLE}.pacbio-blasr-grch38-reheader.20180102.chr*.bam"
+		for F in ${SAMPLE}.pacbio-blasr-grch38-reheader.20180102.*.bam
+		do
+		    if [ ! -f ${F}.bai ]
+		    then
+			samtools index ${F}
+		    fi
+		done
+		samtools merge ${SAMPLE}.pacbio.bam ${SAMPLE}.pacbio-blasr-grch38-reheader.20180102.*.bam
+		samtools index ${SAMPLE}.pacbio.bam
+		rm ${SAMPLE}.pacbio-blasr-grch38-reheader.20180102.chr*
+	    fi
+	    ${BASEDIR}/../src/alfred qc -r GRCh38_full_analysis_set_plus_decoy_hla.fa -f json -o ${SAMPLE}.dna.pacbio.se.json.gz ${SAMPLE}.pacbio.bam
+	done
+	mv ${SAMPLE}.dna.pacbio.se.json.gz dna.pacbio.se.ms.json.gz 
+    fi
     
     # Hi-C
-    for SAMPLE in HG00732 HG00733
-    do
-	if [ ! -f ${SAMPLE}_Hi-C_biorep2_merged_filtered.bam ]
-	then
-	    wget "ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/data_collections/hgsv_sv_discovery/working/20160822_HiC_bam_files/${SAMPLE}_Hi-C_biorep2_merged_filtered.bam"
-	    samtools index ${SAMPLE}_Hi-C_biorep2_merged_filtered.bam
-	fi
-    done
+    if [ ! -f hic.illumina.pe.ms.json.gz ]
+    then
+	for SAMPLE in HG00732 HG00733
+	do
+	    if [ ! -f ${SAMPLE}_Hi-C_biorep2_merged_filtered.bam ]
+	    then
+		wget "ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/data_collections/hgsv_sv_discovery/working/20160822_HiC_bam_files/${SAMPLE}_Hi-C_biorep2_merged_filtered.bam"
+		samtools index ${SAMPLE}_Hi-C_biorep2_merged_filtered.bam
+	    fi
+	    ${BASEDIR}/../src/alfred qc -r GRCh38_full_analysis_set_plus_decoy_hla.fa -f json -o ${SAMPLE}.hic.illumina.pe.json.gz ${SAMPLE}_Hi-C_biorep2_merged_filtered.bam
+	done
+	python ${BASEDIR}/../scripts/merge.py *.hic.illumina.pe.json.gz | gzip -c > hic.illumina.pe.ms.json.gz
+	rm *.hic.illumina.pe.json.gz
+    fi
 else
     echo "Unknown mode ${1}"
 fi
